feature/dynamic-column-width: refactoring in header-component. Calculation now done in table component and passed down as Input to header. This way it can be re-used for the rows
