+++ apps/ngx-table/src/app/table/table.component.ts
@@ -20,6 +20,8 @@ import { DataColumn } from './models/data-column.model';
 import { takeUntil, tap, map } from 'rxjs/operators';
 import { HiddenColumns } from './models/hidden-column.model';
 import { getRowStyle } from './config/row-style-maps';
+import { calcAdjustedWidths } from '../shared/util/calculate-normalized-widths';
+import { parseUnit } from '../shared/util/parse-unit';
 
 @Component({
   selector: 'ngx-table-table',
@@ -39,6 +41,9 @@ export class TableComponent<T> implements OnInit, TableBehavior, OnDestroy {
   @Input() set headerDefinition(headerDefinition: TitleColumn[]) {
     if (headerDefinition) {
       this.stateService.setHeaderDefinition(headerDefinition);
+      // const visible: TitleColumn[] = headerDefinition.filter(v => !v.hide);
+      // console.log(visible);
+      // this.displayColumnWidths$$.next(this.getColumnWidths(visible));
     }
   }
 
@@ -83,9 +88,16 @@ export class TableComponent<T> implements OnInit, TableBehavior, OnDestroy {
 
   public columnWidth: string;
 
+  private displayColumnWidths$$ = new Subject();
+  public displayColumnWidths$: Observable<string>;
+
   constructor(public stateService: TableStateService<T>) {}
 
   ngOnInit() {
+    this.displayColumnWidths$ = this.stateService.renderHeaders$.pipe(
+      map(header => this.getColumnWidths(header))
+    );
+    this.displayColumnWidths$.subscribe(console.log);
     this.renderHeaders$ = this.stateService.renderHeaders$;
     this.renderRows$ = this.stateService.renderRows$;
     this.renderColumnCount$ = this.stateService.renderColumnCount$;
@@ -107,6 +119,12 @@ export class TableComponent<T> implements OnInit, TableBehavior, OnDestroy {
       )
       .subscribe();
     this.columnWidth = this.initializeColumnWidth(this._tableConfig);
+
+    // this.renderHeaders$
+    //   .pipe(takeUntil(this.destroy$))
+    //   .subscribe(header =>
+    //     this.displayColumnWidths$$.next(this.getColumnWidths(header))
+    //   );
   }
   ngOnDestroy() {
     this.destroy$.next();
@@ -154,5 +172,39 @@ export class TableComponent<T> implements OnInit, TableBehavior, OnDestroy {
     } else {
       return DEFAULT_TABLE_CONFIG.defaultColumnWidth;
     }
+  }
+
+  getColumnWidths(displayColumns: TitleColumn[]): string {
+    let columnWidths: string[];
+    let widthUnit: string;
+    let normalizedWidths: string[];
+    if (displayColumns) {
+      columnWidths = displayColumns.map(v => {
+        if (v.width) {
+          return v.width;
+        } else {
+          return DEFAULT_TABLE_CONFIG.defaultColumnWidth;
+        }
+      });
+      // TODO rethink this. It will not be really safe with not %-values.
+      const widths: number[] = displayColumns.map(v => {
+        if (v.width) {
+          widthUnit = parseUnit(v.width);
+          return parseInt(v.width.split(widthUnit)[0]);
+        }
+      });
+
+      const sum = widths.reduce((a, b) => a + b, 0);
+      if (sum > 100) {
+        console.warn('The specified column widths exceed 100%');
+      }
+      /**
+       * Calculate normalized widths when widts are given with '%'
+       * to still reach 100 %.
+       */
+      normalizedWidths = calcAdjustedWidths(widths);
+    }
+
+    return normalizedWidths.join(' ');
   }
 }
